+++ src/routes/__root.tsx
@@ -5,14 +5,8 @@ import { TanStackRouterDevtoolsPanel } from "@tanstack/react-router-devtools";
 import TanStackQueryDevtools from "../integrations/tanstack-query/devtools";
 
 import { BottomNavigation } from "@/components/bottom-navigation";
-import { StationDrawer } from "@/components/maps/station-drawer";
-import { useOverlay } from "@/hooks/use-overlay";
+import { StationMap } from "@/components/maps/station-map";
 import type { QueryClient } from "@tanstack/react-query";
-import {
-	APIProvider,
-	AdvancedMarker,
-	Map as GoogleMap,
-} from "@vis.gl/react-google-maps";
 
 interface MyRouterContext {
 	queryClient: QueryClient;
@@ -20,34 +14,12 @@ interface MyRouterContext {
 
 export const Route = createRootRouteWithContext<MyRouterContext>()({
 	component: () => {
-		const overlay = useOverlay();
-		const position = { lat: 53.54992, lng: 10.00678 };
-
 		return (
 			<>
 				<div className="min-h-screen bg-gray-100">
 					<div className="max-w-screen-sm mx-auto">
 						<Outlet />
-
-						<APIProvider apiKey={import.meta.env.VITE_GOOGLE_MAPS_API_KEY}>
-							<GoogleMap
-								style={{ width: "100%", height: "100vh" }}
-								defaultCenter={position}
-								defaultZoom={10}
-								mapId="DEMO_MAP_ID"
-								disableDefaultUI={true}
-								renderingType="VECTOR"
-							>
-								<AdvancedMarker
-									position={position}
-									onClick={() =>
-										overlay.open(({ isOpen, close }) => (
-											<StationDrawer open={isOpen} close={close} />
-										))
-									}
-								/>
-							</GoogleMap>
-						</APIProvider>
+						<StationMap />
 						<BottomNavigation />
 					</div>
 				</div>
